# Add ICollection<T> shortcut for Seq.contains