I noticed a performance difference between Linq's `Enumerable.Contains` and F#'s `Seq.contains`. Linq uses a test for `ICollection<T>` and if possible defers to `ICollection<T>.Contains`.

https://github.com/dotnet/runtime/blob/d019e70d2b7c2f7cd1137fac084dbcdc3d2e05f5/src/libraries/System.Linq/src/System/Linq/Contains.cs#L11-L12

The code change here introduces a similar shortcut to F#'s `Seq.contains`.
https://github.com/Jlobblet/fsharp/blob/9cb72dea06f52da7bfc682d3e9bb38a42b69ad29/src/fsharp/FSharp.Core/seq.fs#L535-L545

This generally has a performance gain for things that implement `ICollection<T>`, since they typically have faster ways of testing for whether they contain something than enumerating the entire collection.

I have prepared a BenchmarkDotNet benchmark suite to test the performance of this change over a variety of cases. It is available at https://github.com/Jlobblet/fsharp-seq-contains.

The following table compares performance of the existing `Seq.contains`, the version of `Seq.contains` proposed by this PR, and Linq's `Enumerable.Contains`.
Method is which of these methods was used to in the benchmark.
Init is what type of collection was used.
N is how many elements are in the collection. Collections are created from the range `1 .. N`.
containsBehaviour is what target value is passed to `Contains`. It is determined as follows:
```f#
this.target <-
    match this.containsBehaviour with
    | ContainsBehaviour.Start -> 1
    | ContainsBehaviour.Middle -> this.N / 2
    | ContainsBehaviour.End -> this.N
    | ContainsBehaviour.NotPresent -> 0
```

<details>
  <summary>Table of results</summary>
  
  ``` ini

BenchmarkDotNet=v0.13.0, OS=Windows 10.0.19042.1110 (20H2/October2020Update)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET SDK=5.0.302
  [Host]     : .NET 5.0.8 (5.0.821.31504), X64 RyuJIT DEBUG
  DefaultJob : .NET 5.0.8 (5.0.821.31504), X64 RyuJIT


```
|   Method |       Init |      N | containsBehaviour |             Mean |          Error |          StdDev |           Median |
|--------- |----------- |------- |------------------ |-----------------:|---------------:|----------------:|-----------------:|
|  **Current** |   **F# Array** |      **0** |             **Start** |        **24.572 ns** |      **0.4519 ns** |       **0.7425 ns** |        **24.561 ns** |
| Proposed |   F# Array |      0 |             Start |        17.425 ns |      0.3787 ns |       0.3163 ns |        17.394 ns |
|     Linq |   F# Array |      0 |             Start |        19.053 ns |      0.4007 ns |       0.7721 ns |        18.926 ns |
|  **Current** |   **F# Array** |      **0** |            **Middle** |        **20.946 ns** |      **0.4620 ns** |       **1.3549 ns** |        **20.602 ns** |
| Proposed |   F# Array |      0 |            Middle |        17.264 ns |      0.3830 ns |       0.6808 ns |        17.471 ns |
|     Linq |   F# Array |      0 |            Middle |        17.986 ns |      0.3976 ns |       0.6307 ns |        18.017 ns |
|  **Current** |   **F# Array** |      **0** |               **End** |        **20.681 ns** |      **0.4469 ns** |       **0.8610 ns** |        **20.873 ns** |
| Proposed |   F# Array |      0 |               End |        17.099 ns |      0.3773 ns |       0.6406 ns |        17.132 ns |
|     Linq |   F# Array |      0 |               End |        17.662 ns |      0.3689 ns |       0.3623 ns |        17.758 ns |
|  **Current** |   **F# Array** |      **0** |        **NotPresent** |        **20.251 ns** |      **0.4364 ns** |       **0.9579 ns** |        **20.355 ns** |
| Proposed |   F# Array |      0 |        NotPresent |        17.205 ns |      0.3819 ns |       0.5477 ns |        17.303 ns |
|     Linq |   F# Array |      0 |        NotPresent |        17.477 ns |      0.3786 ns |       0.4208 ns |        17.611 ns |
|  **Current** |   **F# Array** |      **1** |             **Start** |        **25.612 ns** |      **0.5452 ns** |       **0.8489 ns** |        **25.490 ns** |
| Proposed |   F# Array |      1 |             Start |        17.951 ns |      0.3521 ns |       0.5481 ns |        18.005 ns |
|     Linq |   F# Array |      1 |             Start |        18.473 ns |      0.3920 ns |       0.3850 ns |        18.437 ns |
|  **Current** |   **F# Array** |      **1** |            **Middle** |        **19.787 ns** |      **0.4274 ns** |       **0.7022 ns** |        **19.608 ns** |
| Proposed |   F# Array |      1 |            Middle |        16.898 ns |      0.3542 ns |       0.4079 ns |        16.761 ns |
|     Linq |   F# Array |      1 |            Middle |        17.578 ns |      0.3853 ns |       0.3956 ns |        17.590 ns |
|  **Current** |   **F# Array** |      **1** |               **End** |        **25.298 ns** |      **0.5259 ns** |       **0.4106 ns** |        **25.429 ns** |
| Proposed |   F# Array |      1 |               End |        17.748 ns |      0.3719 ns |       0.3479 ns |        17.801 ns |
|     Linq |   F# Array |      1 |               End |        18.808 ns |      0.3592 ns |       0.3184 ns |        18.861 ns |
|  **Current** |   **F# Array** |      **1** |        **NotPresent** |        **19.932 ns** |      **0.4202 ns** |       **0.5752 ns** |        **19.841 ns** |
| Proposed |   F# Array |      1 |        NotPresent |        16.741 ns |      0.3678 ns |       0.5276 ns |        16.651 ns |
|     Linq |   F# Array |      1 |        NotPresent |        17.894 ns |      0.3668 ns |       0.3431 ns |        17.948 ns |
|  **Current** |   **F# Array** |     **10** |             **Start** |        **26.404 ns** |      **0.5575 ns** |       **0.7632 ns** |        **26.355 ns** |
| Proposed |   F# Array |     10 |             Start |        15.632 ns |      0.3426 ns |       0.3204 ns |        15.753 ns |
|     Linq |   F# Array |     10 |             Start |        16.385 ns |      0.3567 ns |       0.3663 ns |        16.365 ns |
|  **Current** |   **F# Array** |     **10** |            **Middle** |        **47.995 ns** |      **0.9966 ns** |       **1.4916 ns** |        **47.784 ns** |
| Proposed |   F# Array |     10 |            Middle |        16.125 ns |      0.3466 ns |       0.4507 ns |        15.950 ns |
|     Linq |   F# Array |     10 |            Middle |        16.934 ns |      0.3734 ns |       0.3493 ns |        16.998 ns |
|  **Current** |   **F# Array** |     **10** |               **End** |        **75.839 ns** |      **1.5313 ns** |       **2.3384 ns** |        **76.171 ns** |
| Proposed |   F# Array |     10 |               End |        19.811 ns |      0.4052 ns |       0.4161 ns |        19.714 ns |
|     Linq |   F# Array |     10 |               End |        20.187 ns |      0.3920 ns |       0.3475 ns |        20.216 ns |
|  **Current** |   **F# Array** |     **10** |        **NotPresent** |        **20.387 ns** |      **0.4310 ns** |       **0.6181 ns** |        **20.452 ns** |
| Proposed |   F# Array |     10 |        NotPresent |        15.276 ns |      0.3334 ns |       0.3568 ns |        15.412 ns |
|     Linq |   F# Array |     10 |        NotPresent |        15.892 ns |      0.3227 ns |       0.3019 ns |        15.837 ns |
|  **Current** |   **F# Array** |   **1000** |             **Start** |        **25.179 ns** |      **0.5329 ns** |       **0.6929 ns** |        **25.150 ns** |
| Proposed |   F# Array |   1000 |             Start |        15.653 ns |      0.2900 ns |       0.2571 ns |        15.713 ns |
|     Linq |   F# Array |   1000 |             Start |        16.453 ns |      0.3634 ns |       0.4185 ns |        16.440 ns |
|  **Current** |   **F# Array** |   **1000** |            **Middle** |     **2,763.895 ns** |     **55.2321 ns** |      **77.4277 ns** |     **2,736.230 ns** |
| Proposed |   F# Array |   1000 |            Middle |       156.258 ns |      3.1471 ns |       3.9801 ns |       156.645 ns |
|     Linq |   F# Array |   1000 |            Middle |       151.890 ns |      2.4161 ns |       2.2600 ns |       152.423 ns |
|  **Current** |   **F# Array** |   **1000** |               **End** |     **5,508.473 ns** |    **109.7692 ns** |     **126.4104 ns** |     **5,492.419 ns** |
| Proposed |   F# Array |   1000 |               End |       307.050 ns |      6.0648 ns |       7.4481 ns |       309.572 ns |
|     Linq |   F# Array |   1000 |               End |       313.104 ns |      5.4332 ns |       4.8164 ns |       313.563 ns |
|  **Current** |   **F# Array** |   **1000** |        **NotPresent** |        **19.914 ns** |      **0.4232 ns** |       **0.4703 ns** |        **19.924 ns** |
| Proposed |   F# Array |   1000 |        NotPresent |        14.943 ns |      0.3037 ns |       0.2692 ns |        14.850 ns |
|     Linq |   F# Array |   1000 |        NotPresent |        16.495 ns |      0.3416 ns |       0.3355 ns |        16.612 ns |
|  **Current** |   **F# Array** | **100000** |             **Start** |        **26.436 ns** |      **0.5460 ns** |       **0.7654 ns** |        **26.430 ns** |
| Proposed |   F# Array | 100000 |             Start |        16.164 ns |      0.2412 ns |       0.2256 ns |        16.234 ns |
|     Linq |   F# Array | 100000 |             Start |        16.527 ns |      0.2116 ns |       0.1875 ns |        16.566 ns |
|  **Current** |   **F# Array** | **100000** |            **Middle** |   **275,900.021 ns** |  **5,466.2909 ns** |   **7,482.3207 ns** |   **276,312.549 ns** |
| Proposed |   F# Array | 100000 |            Middle |    13,547.143 ns |    235.4926 ns |     208.7580 ns |    13,530.002 ns |
|     Linq |   F# Array | 100000 |            Middle |    13,544.863 ns |    253.2478 ns |     260.0668 ns |    13,560.174 ns |
|  **Current** |   **F# Array** | **100000** |               **End** |   **543,555.491 ns** | **10,769.0041 ns** |  **13,225.3042 ns** |   **546,310.840 ns** |
| Proposed |   F# Array | 100000 |               End |    29,099.498 ns |    573.6598 ns |     804.1909 ns |    28,763.507 ns |
|     Linq |   F# Array | 100000 |               End |    29,091.438 ns |    577.2237 ns |     790.1103 ns |    28,944.771 ns |
|  **Current** |   **F# Array** | **100000** |        **NotPresent** |        **20.024 ns** |      **0.4169 ns** |       **0.4094 ns** |        **20.180 ns** |
| Proposed |   F# Array | 100000 |        NotPresent |        15.773 ns |      0.2514 ns |       0.2352 ns |        15.836 ns |
|     Linq |   F# Array | 100000 |        NotPresent |        15.792 ns |      0.3283 ns |       0.3071 ns |        15.861 ns |
|  **Current** |    **F# List** |      **0** |             **Start** |        **26.922 ns** |      **0.4499 ns** |       **0.3989 ns** |        **26.903 ns** |
| Proposed |    F# List |      0 |             Start |        34.319 ns |      0.7227 ns |       0.8323 ns |        34.220 ns |
|     Linq |    F# List |      0 |             Start |        35.105 ns |      0.6679 ns |       0.5920 ns |        35.179 ns |
|  **Current** |    **F# List** |      **0** |            **Middle** |        **23.909 ns** |      **0.4388 ns** |       **0.4105 ns** |        **23.897 ns** |
| Proposed |    F# List |      0 |            Middle |        28.956 ns |      0.6134 ns |       0.7064 ns |        28.873 ns |
|     Linq |    F# List |      0 |            Middle |        31.957 ns |      0.6686 ns |       0.7699 ns |        32.227 ns |
|  **Current** |    **F# List** |      **0** |               **End** |        **22.281 ns** |      **0.4651 ns** |       **0.5712 ns** |        **22.265 ns** |
| Proposed |    F# List |      0 |               End |        28.337 ns |      0.5526 ns |       0.5427 ns |        28.239 ns |
|     Linq |    F# List |      0 |               End |        32.328 ns |      0.6226 ns |       0.8522 ns |        32.256 ns |
|  **Current** |    **F# List** |      **0** |        **NotPresent** |        **23.763 ns** |      **0.5019 ns** |       **0.4930 ns** |        **23.843 ns** |
| Proposed |    F# List |      0 |        NotPresent |        28.678 ns |      0.6030 ns |       0.6452 ns |        28.638 ns |
|     Linq |    F# List |      0 |        NotPresent |        31.719 ns |      0.5911 ns |       0.5529 ns |        31.803 ns |
|  **Current** |    **F# List** |      **1** |             **Start** |        **31.324 ns** |      **0.6480 ns** |       **0.9294 ns** |        **31.338 ns** |
| Proposed |    F# List |      1 |             Start |        36.183 ns |      0.7499 ns |       0.9209 ns |        35.997 ns |
|     Linq |    F# List |      1 |             Start |        40.374 ns |      0.8016 ns |       1.7082 ns |        39.872 ns |
|  **Current** |    **F# List** |      **1** |            **Middle** |        **24.466 ns** |      **0.5260 ns** |       **1.3388 ns** |        **24.388 ns** |
| Proposed |    F# List |      1 |            Middle |        31.057 ns |      0.6537 ns |       1.8005 ns |        31.162 ns |
|     Linq |    F# List |      1 |            Middle |        33.843 ns |      0.6947 ns |       1.4653 ns |        33.849 ns |
|  **Current** |    **F# List** |      **1** |               **End** |        **31.899 ns** |      **0.6731 ns** |       **1.1965 ns** |        **31.802 ns** |
| Proposed |    F# List |      1 |               End |        39.889 ns |      0.8280 ns |       1.9996 ns |        39.693 ns |
|     Linq |    F# List |      1 |               End |        42.709 ns |      0.8677 ns |       1.2987 ns |        42.567 ns |
|  **Current** |    **F# List** |      **1** |        **NotPresent** |        **24.053 ns** |      **0.4998 ns** |       **0.4908 ns** |        **23.966 ns** |
| Proposed |    F# List |      1 |        NotPresent |        30.430 ns |      0.6125 ns |       0.5730 ns |        30.440 ns |
|     Linq |    F# List |      1 |        NotPresent |        33.790 ns |      0.5517 ns |       0.4891 ns |        33.869 ns |
|  **Current** |    **F# List** |     **10** |             **Start** |        **32.577 ns** |      **0.6900 ns** |       **1.2617 ns** |        **32.202 ns** |
| Proposed |    F# List |     10 |             Start |        38.131 ns |      0.7946 ns |       1.7773 ns |        37.763 ns |
|     Linq |    F# List |     10 |             Start |        42.965 ns |      0.8983 ns |       1.4248 ns |        43.182 ns |
|  **Current** |    **F# List** |     **10** |            **Middle** |        **60.606 ns** |      **1.2320 ns** |       **2.5716 ns** |        **60.787 ns** |
| Proposed |    F# List |     10 |            Middle |        66.290 ns |      1.3644 ns |       2.3536 ns |        66.176 ns |
|     Linq |    F# List |     10 |            Middle |        85.968 ns |      1.9886 ns |       5.3424 ns |        85.394 ns |
|  **Current** |    **F# List** |     **10** |               **End** |       **118.409 ns** |      **2.3945 ns** |       **2.9406 ns** |       **118.205 ns** |
| Proposed |    F# List |     10 |               End |       128.567 ns |      2.5719 ns |       3.4334 ns |       128.805 ns |
|     Linq |    F# List |     10 |               End |       130.990 ns |      2.6419 ns |       3.3411 ns |       129.827 ns |
|  **Current** |    **F# List** |     **10** |        **NotPresent** |        **33.645 ns** |      **2.5735 ns** |       **7.5476 ns** |        **29.963 ns** |
| Proposed |    F# List |     10 |        NotPresent |        32.368 ns |      0.6887 ns |       1.7655 ns |        32.204 ns |
|     Linq |    F# List |     10 |        NotPresent |        35.122 ns |      0.8751 ns |       2.5108 ns |        34.625 ns |
|  **Current** |    **F# List** |   **1000** |             **Start** |        **34.667 ns** |      **0.8245 ns** |       **2.3920 ns** |        **33.976 ns** |
| Proposed |    F# List |   1000 |             Start |        40.166 ns |      0.8154 ns |       1.3167 ns |        40.180 ns |
|     Linq |    F# List |   1000 |             Start |        45.019 ns |      0.9225 ns |       1.5412 ns |        44.790 ns |
|  **Current** |    **F# List** |   **1000** |            **Middle** |     **4,177.541 ns** |     **82.8319 ns** |     **163.5021 ns** |     **4,134.399 ns** |
| Proposed |    F# List |   1000 |            Middle |     4,070.790 ns |     81.3350 ns |     138.1131 ns |     4,078.710 ns |
|     Linq |    F# List |   1000 |            Middle |     4,126.259 ns |     81.7793 ns |     149.5380 ns |     4,104.058 ns |
|  **Current** |    **F# List** |   **1000** |               **End** |     **7,879.427 ns** |    **155.3649 ns** |     **272.1091 ns** |     **7,817.849 ns** |
| Proposed |    F# List |   1000 |               End |     8,401.951 ns |    167.5125 ns |     306.3060 ns |     8,388.686 ns |
|     Linq |    F# List |   1000 |               End |     8,197.380 ns |    153.8380 ns |     285.1487 ns |     8,118.596 ns |
|  **Current** |    **F# List** |   **1000** |        **NotPresent** |        **26.373 ns** |      **0.5672 ns** |       **0.9319 ns** |        **26.407 ns** |
| Proposed |    F# List |   1000 |        NotPresent |        32.830 ns |      0.6851 ns |       1.5741 ns |        32.456 ns |
|     Linq |    F# List |   1000 |        NotPresent |        35.827 ns |      0.7563 ns |       1.9522 ns |        35.618 ns |
|  **Current** |    **F# List** | **100000** |             **Start** |        **34.569 ns** |      **0.7322 ns** |       **1.6823 ns** |        **34.496 ns** |
| Proposed |    F# List | 100000 |             Start |        41.888 ns |      0.8596 ns |       2.2189 ns |        41.291 ns |
|     Linq |    F# List | 100000 |             Start |        46.756 ns |      0.8211 ns |       0.7681 ns |        46.833 ns |
|  **Current** |    **F# List** | **100000** |            **Middle** |   **412,063.874 ns** |  **8,032.0563 ns** |  **11,259.8211 ns** |   **410,256.055 ns** |
| Proposed |    F# List | 100000 |            Middle |   429,387.141 ns |  8,267.9596 ns |  12,872.2170 ns |   432,189.453 ns |
|     Linq |    F# List | 100000 |            Middle |   388,206.417 ns |  7,574.5604 ns |  11,102.6924 ns |   389,472.852 ns |
|  **Current** |    **F# List** | **100000** |               **End** |   **877,624.223 ns** | **16,290.1187 ns** |  **38,715.1906 ns** |   **871,745.215 ns** |
| Proposed |    F# List | 100000 |               End |   873,843.478 ns | 16,880.4026 ns |  26,280.7532 ns |   878,306.836 ns |
|     Linq |    F# List | 100000 |               End |   806,492.687 ns | 16,001.9066 ns |  17,121.8607 ns |   808,374.414 ns |
|  **Current** |    **F# List** | **100000** |        **NotPresent** |        **25.854 ns** |      **0.5454 ns** |       **0.5601 ns** |        **25.773 ns** |
| Proposed |    F# List | 100000 |        NotPresent |        33.085 ns |      0.7016 ns |       1.2830 ns |        32.920 ns |
|     Linq |    F# List | 100000 |        NotPresent |        36.187 ns |      0.7608 ns |       0.8762 ns |        36.312 ns |
|  **Current** |     **F# Set** |      **0** |             **Start** |        **81.287 ns** |      **1.6075 ns** |       **3.6935 ns** |        **81.296 ns** |
| Proposed |     F# Set |      0 |             Start |        13.053 ns |      0.2905 ns |       0.4773 ns |        13.029 ns |
|     Linq |     F# Set |      0 |             Start |        13.046 ns |      0.2921 ns |       0.4372 ns |        12.973 ns |
|  **Current** |     **F# Set** |      **0** |            **Middle** |        **64.977 ns** |      **1.3301 ns** |       **3.7734 ns** |        **64.504 ns** |
| Proposed |     F# Set |      0 |            Middle |        12.280 ns |      0.2806 ns |       0.4025 ns |        12.303 ns |
|     Linq |     F# Set |      0 |            Middle |        12.132 ns |      0.2781 ns |       0.3415 ns |        12.165 ns |
…