@Jlobblet 2021-07-25
This test failure is down to how the `contains` function is implemented in different modules. For example, in `Array.contains`:
```f#
[<CompiledName("Contains")>]
let inline contains value (array: 'T[]) =
    checkNonNull "array" array
    let mutable state = false
    let mutable i = 0
    while not state && i < array.Length do
        state <- value = array.[i]
        i <- i + 1
    state
```
And since `nan <> nan`, this returns `false`.

By comparison, a BCL array's implementation of `Contains` evenually boils down to using `.Equals`: https://github.com/dotnet/runtime/blob/ce3e00b9d5e0135af68fa878f48952a52614e0ef/src/coreclr/System.Private.CoreLib/src/System/Collections/Generic/EqualityComparer.CoreCLR.cs#L33

Essentially, this is the difference between the two following cases:
```f#
[| nan = nan
   nan.Equals nan |]
|> Array.iter (printfn "%b")
```

I'm not sure what the best way to continue here is. I'll wait to see what people think.