@dsyme 2021-07-26
> I'm not sure what the best way to continue here is. I'll wait to see what people think.

The improvement for ICollection will be because iteration is faster.  The best way forward would be to use specific iteration code for ICollection in F# but make sure you use the right `=` comparison.

Note that all this code is inlined (to get type-specialization on the `=`), which might mean we assess the contribution according to code size implications - all the ICollection code will also get inlined for each use of `Seq.contains` which might be a bit much