@Jlobblet 2021-07-27
> The improvement for ICollection will be because iteration is faster. The best way forward would be to use specific iteration code for ICollection in F# but make sure you use the right `=` comparison.

I think the gain is from the data structures themselves performing more efficient iteration (or otherwise) by implementing specialised versions of Contains, rather than anything especially unique to ICollection. Iterating over an ICollection is the same as an IEnumerable, if I understand correctly - in fact, its `GetEnumerator` method is inherited from IEnumerable.