@Jlobblet 2021-08-31
I haven't been able to think of a way to use the `ICollection<T>` interface, which delegates implementing `Contains` to the inheritor, to retain the behaviour of `=` and also benefit from a performance increase, so I'll close this PR.